Adds spec functions for unvisited node logic
Adds specification functions related to finding the minimum unvisited node in Dijkstra's algorithm.
These specifications include functions to determine if there are unvisited nodes, if a node is a candidate for the minimum,
and a core recursive function to find the minimum unvisited node.
Proofs still need to be added.